refactor: MainActivity 클래스 navui 패키지로 이동
