+++ app/src/main/AndroidManifest.xml
@@ -39,7 +39,7 @@
             android:name=".ui.write.WriteActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.main.MainActivity"
+            android:name=".ui.navui.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
