+++ app/src/main/java/com/moroom/android/ui/login/LoginActivity.java
@@ -12,7 +12,7 @@ import android.widget.Toast;
 import com.moroom.android.R;
 import com.moroom.android.databinding.ActivityLoginBinding;
 import com.google.firebase.auth.FirebaseAuth;
-import com.moroom.android.ui.main.MainActivity;
+import com.moroom.android.ui.navui.MainActivity;
 import com.moroom.android.ui.signup.SignupActivity;
 
 public class LoginActivity extends AppCompatActivity {
