+++ app/src/main/java/com/moroom/android/ui/signup/SignupActivity.java
@@ -23,7 +23,7 @@ import com.google.firebase.auth.FirebaseUser;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.dynamiclinks.FirebaseDynamicLinks;
-import com.moroom.android.ui.main.MainActivity;
+import com.moroom.android.ui.navui.MainActivity;
 
 public class SignupActivity extends AppCompatActivity {
     private ActivitySignUpBinding binding;
