+++ app/src/main/res/layout/fragment_profile.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:gravity="center_vertical"
     android:orientation="vertical"
-    tools:context=".ui.main.MainActivity">
+    tools:context=".ui.navui.MainActivity">
 
     <LinearLayout
         android:id="@+id/linearTop"
